Fix typescript external dependency
What: Update path to tsc in npm scripts

Why: When initially writing this code, I had assumed that the path I provided
to the typescript compiler in the npm scripts was referencing the locally
installed version of typescript. I was wrong. It still references the
globally installed version.

This presents two major issues:

- Users will not be able to install / run without typescript pre-installed
  globally in either npm or yarn.
- Not using a pinned local version of typescript opens this to future
  incompatibility breakage.